requirements.txt
discord.py==2.3.1
Pillow==10.0.0
Requests==2.31.0
